requirements.txt
pytest
aws-cdk-lib
cfn-lint
constructs
syrupy